+++ source/qiti_FunctionCallData_Impl.hpp
@@ -34,10 +34,10 @@ struct FunctionCallData::Impl
     std::chrono::steady_clock::time_point end_time;
     std::thread::id callingThread;
     
-    uint timeSpentInFunctionNanoseconds = 0;
+    uint64_t timeSpentInFunctionNanoseconds = 0;
     
-    uint numHeapAllocationsBeforeFunctionCall = 0;
-    uint numHeapAllocationsAfterFunctionCall  = 0;
+    uint64_t numHeapAllocationsBeforeFunctionCall = 0;
+    uint64_t numHeapAllocationsAfterFunctionCall  = 0;
 };
 } // namespace qiti
